+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 
 ## [unreleased]
 
+ * HTTP Proxy: Allow it to intercept/modify the response from the target service.
+   **BREAKING CHANGE**: The *MashroomHttpProxyInterceptor* interface changed and is not compatible with the old one:
+      * intercept() has been renamed to interceptRequest()
+      * A new method interceptResponse() has been added
+      * Instead of returning *result.reject* you can now call res.sendStatus() yourself and just return *result.responseHandled* as
+        a hint that the proxy doesn't have to do anything.
  * Portal: Fixed the client log handler (didn't terminate correctly)
 
 ## 1.6.3 (December 20, 2020)